Don't recurse on files, just the sub-dirs
We'll hit a 5K GitHub API limit (per hour) if we update more than a few
times per hour, or if the number of projects in jsDelivr goes up. The
update is slow now (could increase the limit on async.eachLimit, or
lower the GitHub timeout), but it works for now.

A commit hook on the jsdelivr repo might work better to build the JSON
then, rather than having api-sync build it on-demand.